+++ src/clj_kondo/impl/analyzer.clj
@@ -1840,11 +1840,17 @@
   (when-let [children (next (:children expr))]
     (let [[method-name-node dispatch-val-node & fn-tail] children
           ctx-without-idx (dissoc ctx :idx :len)
+          dispatch-val-str (pr-str (sexpr dispatch-val-node))
+          method-name (sexpr method-name-node)
+          ns-name (-> ctx :ns :name)
           _ (analyze-usages2 (assoc ctx-without-idx
                                     :defmethod true,
-                                    :dispatch-val-str (pr-str (sexpr dispatch-val-node)))
+                                    :dispatch-val-str dispatch-val-str)
                              method-name-node)
-          _ (analyze-expression** ctx-without-idx dispatch-val-node)]
+          _ (analyze-expression** ctx-without-idx dispatch-val-node)
+          _ (when (and method-name ns-name)
+              (namespace/reg-defmethod! ctx-without-idx ns-name method-name 
+                                        dispatch-val-str expr))]
       (analyze-fn ctx-without-idx (with-meta {:children (cons nil fn-tail)} (meta expr))))))
 
 (defn analyze-areduce [ctx expr]
@@ -2285,7 +2291,7 @@
                 (swap! nss update-in [base-lang lang ns-name]
                        (fn [ns]
                          (-> ns
-                             (update :clojure-excluded (fnil conj #{}) 
+                             (update :clojure-excluded (fnil conj #{})
                                      (with-meta sym excluded-meta))
                              (update :vars dissoc sym)
                              (update :var-counts dissoc sym))))))))))