+++ CHANGELOG.md
@@ -25,6 +25,7 @@ For a list of breaking changes, check [here](#breaking-changes).
 
 ## 2026.01.19
 
+- [#2528](https://github.com/clj-kondo/clj-kondo/issues/2528): NEW linter: `:shadowed-defmethod` which warns when the same `defmethod` multimethod + dispatch value is defined multiple times.([@jramosg](https://github.com/jramosg))
 - [#2735](https://github.com/clj-kondo/clj-kondo/issues/2735): NEW linter: `:duplicate-refer` which warns on duplicate entries in `:refer` of `:require`. ([@jramosg](https://github.com/jramosg))
 - [#2734](https://github.com/clj-kondo/clj-kondo/issues/2734): NEW linter: `:aliased-referred-var`, which warns when a var is both referred and accessed via an alias in the same namespace. ([@jramosg](https://github.com/jramosg))
 - [#2745](https://github.com/clj-kondo/clj-kondo/issues/2745): NEW linter: `:is-message-not-string` which warns when `clojure.test/is` receives a non-string message argument. This linter replaces the previous type-mismatch enforcement for `is` message arguments and can be disabled to allow non-string values. ([@jramosg](https://github.com/jramosg))